refactor: move the scene into rav-appearance, leaving the rasteriser alone
Scene, Band, Style, StyleId and CapStyle are a description of a frame, not a way
of drawing one. They sat in the same file as tiny-skia, which meant the seam
between the analyser and every surface was defined inside the pixel surface.

Drawing becomes a `Draw` trait implemented in rav for `Scene`. An extension
trait rather than an inherent method because rav-appearance must never name a
rasteriser: a scene that knew how to rasterise itself would be one only a
rasteriser could consume, and an LED matrix has none. Call sites still read
`scene.draw(&mut canvas)`.

src/render/raster.rs is now what its name says - a Canvas and the one impl that
puts a scene onto it.
