+++ crates/rav-appearance/src/lib.rs
@@ -39,6 +39,8 @@ extern crate std;
 
 pub mod ink;
 pub mod ramp;
+pub mod scene;
 
 pub use ink::{ANSI_NAMES, Colour, DEFAULT_ANSI, Ink};
 pub use ramp::{Ramp, Stripe};
+pub use scene::{Band, CapStyle, Scene, Style, StyleId};
